requirements.txt
llama_cpp_python==0.2.65
tokenizers==0.15.2
transformers==4.39.3
streamlit==1.34.0
graphviz==0.20.3
rich==13.7.1